adicionando funções no componente

+++ src/app/services/user.service.ts
@@ -8,15 +8,18 @@ export class UserService {
   apiUrl:string='http://localhost:8080/user'
   constructor(private httpClient:HttpClient) { }
 
-  edit(email:string,name:string,password:string){
+  edit(email:string|null,name:string,password:string){
     return this.httpClient.put<void>(this.apiUrl,{
       email,
       name,
       password
     })
   }
-  delete(email:string){
-    const params = new HttpParams().set('email',email)
-    return this.httpClient.delete<void>(this.apiUrl,{params})
+  delete(email:string|null){
+    if(email){
+      const params = new HttpParams().set('email',email)
+      return this.httpClient.delete<void>(this.apiUrl,{params})
+    }
+    throw new Error
   }
 }
